pyproject.toml
[tool.poetry]
name = "vkbottle-cm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.6.2"
SQLAlchemy = "^1.4.15"
ipython = "^7.23.1"
python-dotenv = "^0.17.1"
psycopg2 = "^2.8.6"
flake8 = "^3.9.2"
black = "^21.5b1"
oitnb = "^0.2.2"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"